Finalize tooltip code
