+++ specifyweb/frontend/js_src/lib/components/QueryBuilder/Import.tsx
@@ -19,7 +19,6 @@ import { schema } from '../DataModel/schema';
 import type { SpQuery } from '../DataModel/types';
 import { Dialog, LoadingScreen } from '../Molecules/Dialog';
 import { FilePicker, fileToText } from '../Molecules/FilePicker';
-import { QueryFieldSpec } from './fieldSpec';
 
 export function QueryImport({
   onClose: handleClose,
@@ -42,23 +41,8 @@ export function QueryImport({
           acceptedFormats={['.json']}
           onSelected={(file): void =>
             loading(
-              // FIXME: revert
               fileToText(file)
                 .then<SerializedModel<SpQuery>>(f.unary(JSON.parse))
-                .then((query) => {
-                  query.fields.map((field) => {
-                    const fieldSpec = QueryFieldSpec.fromStringId(
-                      field.stringid,
-                      field.isrelfld ?? false
-                    );
-                    const isHidden = fieldSpec.joinPath.some(
-                      ({ isHidden }) => isHidden
-                    );
-                    if (isHidden) return fieldSpec;
-                    else return undefined;
-                  });
-                  return query;
-                })
                 .then(
                   (query) =>
                     new schema.models.SpQuery.Resource(
